[ALPS04140704] scp: mt6885: fixed ipi and clear spm after handler finished
1. enlarge 1 slot size for store id info
2. correct msg_buf pointer to fix memcpy to the wrong place
3. clear spm request after mbox isr finished

MTK-Commit-Id: 7398d8e7fde86051c0cc89c0c623b2d32e7a3298

CR-Id: ALPS04140704
Feature: TinySys

+++ drivers/misc/mediatek/scp/mt6885/scp_ipi.h
@@ -23,10 +23,10 @@ enum scp_ipi_status {
 };
 
 struct scp_ipi_wrapper {
-	uint8_t out_id_0;
-	uint8_t out_id_1;
-	uint8_t in_id_0;
-	uint8_t in_id_1;
+	uint32_t out_id_0;
+	uint32_t out_id_1;
+	uint32_t in_id_0;
+	uint32_t in_id_1;
 	uint32_t out_size;
 	uint32_t in_size;
 	void *msg_0;
